Simplify the navigator, combine into the navigation utilities file
