+++ boilerplate/app/navigation/index.ts
@@ -1,3 +1,3 @@
 export * from "./primary-navigator"
 export * from "./root-navigator"
-export * from "./navigation-service"
+export * from "./navigation-utilities"
